REFACTOR : 오탈자 수정및 코드개선
