+++ src/main/java/com/kindergarten/api/common/result/ResponseService.java
@@ -8,7 +8,7 @@ import java.util.List;
 public class ResponseService {
 
     public enum CommonResponse {
-        SUCCESS(0, "성공하였습니디.");
+        SUCCESS(0, "성공하였습니다.");
 
         int code;
         String msg;
